+++ examples/worker/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @afauthhq/example-worker
 
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @afauthhq/core@0.1.0
+  - @afauthhq/server@0.1.0
+  - @afauthhq/worker@0.1.0
+
 ## 0.0.1-alpha.0
 
 ### Patch Changes